+++ app/src/main/java/com/codesoom/assignment/product/controllers/NotFoundErrorAdvice.java
@@ -1,7 +1,7 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.product.controllers;
 
 import com.codesoom.assignment.ProductNotFoundException;
-import com.codesoom.assignment.dto.ErrorResponse;
+import com.codesoom.assignment.product.dto.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
