+++ app/src/test/java/com/codesoom/assignment/controllers/HelloControllerTest.java
@@ -1,5 +1,6 @@
 package com.codesoom.assignment.controllers;
 
+import com.codesoom.assignment.product.controllers.HelloController;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
